+++ proof-input/src/recursion/mod.rs
@@ -1,2 +1,2 @@
-pub mod uniform;
+pub mod tree_test;
 pub mod leaf_test;
